Switched ALL long words to BigInteger to match the assignment rules pretty awfull to do considering the amount of bitwise operator i used (which had to be converted to biginteger methods) but all tests pass

+++ src/ch/hepia/IL/ReedMuller/NearestSearchFunction.java
@@ -1,7 +1,9 @@
 package ch.hepia.IL.ReedMuller;
 
+import java.math.BigInteger;
+
 public interface NearestSearchFunction {
 
-	int nearestWord(int z, ReedMullerOne rmo);
+	BigInteger nearestWord(BigInteger z, ReedMullerOne rmo);
 	
 }
