+++ src/ch/hepia/IL/ReedMuller/searches/ExaustiveSearch.java
@@ -1,5 +1,7 @@
 package ch.hepia.IL.ReedMuller.searches;
 
+import java.math.BigInteger;
+
 import ch.hepia.IL.ReedMuller.BitUtil;
 import ch.hepia.IL.ReedMuller.NearestSearchFunction;
 import ch.hepia.IL.ReedMuller.ReedMullerOne;
@@ -18,13 +20,13 @@ public class ExaustiveSearch implements NearestSearchFunction {
 	}
 	
 	@Override
-	public int nearestWord(int z, ReedMullerOne rmo) {
-		int nn = 0;
+	public BigInteger nearestWord(BigInteger z, ReedMullerOne rmo) {
+		BigInteger nn = BigInteger.valueOf(0);
 		int nd = Integer.MAX_VALUE;
 		for (int i = 0; i < (1 << rmo.getR()+1); i++) {
-			int dist = BitUtil.HamDist(rmo.encode(i), z);
+			int dist = BitUtil.HamDist(rmo.encode(BigInteger.valueOf(i)), z);
 			if(dist < nd) {
-				nn = i;
+				nn = BigInteger.valueOf(i);
 				nd = dist;
 			}
 		}
